iomap: Prepare page ops for larger I/Os
Just adjust the types for the page_prepare() and page_done() callbacks
to size_t from unsigned int.  While I don't see individual pages that
large in our near future, it's convenient to be able to pass in larger
lengths and has no cost.

+++ fs/gfs2/bmap.c
@@ -1028,7 +1028,7 @@ static void gfs2_write_unlock(struct inode *inode)
 }
 
 static int gfs2_iomap_page_prepare(struct inode *inode, loff_t pos,
-				   unsigned len, struct iomap *iomap)
+				   size_t len, struct iomap *iomap)
 {
 	unsigned int blockmask = i_blocksize(inode) - 1;
 	struct gfs2_sbd *sdp = GFS2_SB(inode);
@@ -1039,7 +1039,7 @@ static int gfs2_iomap_page_prepare(struct inode *inode, loff_t pos,
 }
 
 static void gfs2_iomap_page_done(struct inode *inode, loff_t pos,
-				 unsigned copied, struct page *page,
+				 size_t copied, struct page *page,
 				 struct iomap *iomap)
 {
 	struct gfs2_trans *tr = current->journal_info;
